Reorganized tests
