+++ package.json
@@ -33,5 +33,16 @@
     "docdash": "^0.4.0",
     "jsdoc": "^3.5.5",
     "tui-jsdoc-template": "^1.2.2"
-  }
+  },
+  "ava": {
+		"files": [
+			"tests/unit/**/*.test.js",
+			"!**/not-this-file.js"
+		],
+		"concurrency": 5,
+		"failFast": true,
+		"failWithoutAssertions": true,
+		"tap": false,
+		"powerAssert": false
+	}
 }
