+++ tests/integration/auth.js
