+++ tests/integration/users.js
@@ -3,4 +3,6 @@ const config = require("./config.json");
 
 var client = new Client(config.client);
 
-client.users.fetch().then(console.dir);
+client.users.fetchAll().then(console.dir);
+
+process.on('unhandledRejection', console.error);
